requirements.txt
pandas==2.2.3
numpy==1.26.4
keras==3.6.0
matplotlib==3.10.0
seaborn==0.13.2
umap-learn==0.5.3
scikit-learn==1.6.1
qiskit==1.1.0
qiskit-machine-learning==0.8.2
qiskit-ibm-runtime==0.34.0
tensorflow==2.18.0